requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pygame>=2.5.0
pandas>=2.0.0
networkx>=3.1
Pillow>=10.0.0

